Fix two bf16 footguns in pooled VAE validation path; stage1 patch 96^3
The pilot (weight_dtype=bf16) crashed at the first validation (step 1000) on
two distinct bf16 issues, both surfaced before the real run:

- utils_plot.py / prepare_image_for_logging: validation image logging called
  .cpu().numpy() on a bf16 tensor — numpy has no bfloat16 dtype (fp16 works,
  so the MIUA fp16 path never hit it). Cast to .float() before numpy.
- train_VAE.py best-checkpoint save: guarded the GradScaler state on
  `if args.amp` (True for bf16) but scaler_g/scaler_d are None unless
  use_scaler (amp AND fp16). Switched to `if use_scaler` — matching the
  periodic-checkpoint block and the already-guarded resume path.

stage1 patch 128^3 -> 96^3 (bs/eff/steps unchanged): patch size is not a
fairness lever (applied identically across MAISI/L_SID/L_VAD; 3DMD keeps its
own 64^3), 96^3 is closer to MAISI's own recipe (s1 64^3 -> s2 128^3), and the
pilot measured ~2.3x throughput (compute-bound, no batch-bump needed). stage2
stays 192^3 (inference-scale match).

+++ scripts/utils_plot.py
@@ -96,7 +96,7 @@ def visualize_one_slice_in_3d(image, axis: int = 2, center=None, mask_bool=True,
     if mask_bool:
         draw_img = normalize_label_to_uint8(colorize, draw_img, n_label)
     else:
-        draw_img = draw_img.squeeze().cpu().numpy().astype(np.float32)
+        draw_img = draw_img.squeeze().float().cpu().numpy().astype(np.float32)
         draw_img = np.stack((draw_img,) * 3, axis=-1)
     return draw_img
